[GR-78077] Reshape threaded Crema handlers around fast and slow paths.

+++ substratevm/src/com.oracle.svm.interpreter.metadata/src/com/oracle/svm/interpreter/metadata/ByteUtils.java
@@ -189,7 +189,7 @@ public final class ByteUtils {
         }
     }
 
-    public static void opaqueWrite(byte[] data, int bci, byte value) {
+    public static void opaqueWrite(byte[] data, long bci, byte value) {
         UNSAFE.putByteOpaque(data, offsetFor(bci), value);
     }
 }
